+++ flash/image/detection/model.py
@@ -14,7 +14,7 @@
 from typing import Any, Dict, List, Optional
 
 from flash.core.adapter import AdapterTask
-from flash.core.data.output import Preds
+from flash.core.data.output import PredsOutput
 from flash.core.registry import FlashRegistry
 from flash.core.utilities.types import LR_SCHEDULER_TYPE, OPTIMIZER_TYPE, OUTPUT_TYPE
 from flash.image.detection.backbones import OBJECT_DETECTION_HEADS
@@ -73,7 +73,7 @@ class ObjectDetector(AdapterTask):
             learning_rate=learning_rate,
             optimizer=optimizer,
             lr_scheduler=lr_scheduler,
-            output=output or Preds(),
+            output=output or PredsOutput(),
         )
 
     def _ci_benchmark_fn(self, history: List[Dict[str, Any]]) -> None: